moved enchantment module to item library and added QuiltEnchantableItem
