--- library/content_other/enchantment/build.gradle
@@ -1,18 +1,0 @@
-plugins {
-	id("qsl.module")
-}
-
-qslModule {
-	name = "Quilt Enchantment API"
-	moduleName = "enchantment"
-	id = "quilt_enchantment"
-	description = "An "
-	library = "content_other"
-	moduleDependencies {
-		core {
-			api("qsl_base")
-			testmodOnly("resource_loader")
-		}
-	}
-	accessWidener()
-}
