--- library/content_other/enchantment/src/main/java/org/quiltmc/qsl/enchantment/impl/EnchantmentGodClass.java
@@ -1,26 +1,0 @@
-/*
- * Copyright 2023 QuiltMC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.quiltmc.qsl.enchantment.impl;
-
-import org.jetbrains.annotations.ApiStatus;
-
-import org.quiltmc.qsl.enchantment.api.EnchantingContext;
-
-@ApiStatus.Internal
-public final class EnchantmentGodClass {
-	public static ThreadLocal<EnchantingContext> context = new ThreadLocal<>();
-}
