--- library/content_other/enchantment/src/main/java/org/quiltmc/qsl/enchantment/mixin/EnchantWithLevelsLootFunctionMixin.java
@@ -1,48 +1,0 @@
-/*
- * Copyright 2023 QuiltMC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.quiltmc.qsl.enchantment.mixin;
-
-import org.spongepowered.asm.mixin.Final;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Shadow;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
-
-import net.minecraft.item.ItemStack;
-import net.minecraft.loot.context.LootContext;
-import net.minecraft.loot.function.EnchantWithLevelsLootFunction;
-
-import org.quiltmc.qsl.enchantment.api.EnchantingContext;
-import org.quiltmc.qsl.enchantment.api.QuiltEnchantmentHelper;
-
-@Mixin(EnchantWithLevelsLootFunction.class)
-public class EnchantWithLevelsLootFunctionMixin {
-	@Shadow
-	@Final
-	boolean treasureEnchantmentsAllowed;
-
-	@Inject(method = "process", at = @At("HEAD"))
-	private void setEnchantingLootContext(ItemStack stack, LootContext context, CallbackInfoReturnable<ItemStack> cir) {
-		QuiltEnchantmentHelper.setContext(new EnchantingContext(0, 0, stack, context.getWorld(), context.getRandom(), this.treasureEnchantmentsAllowed));
-	}
-
-	@Inject(method = "process", at = @At("RETURN"))
-	private void removeEnchantingLootContext(ItemStack stack, LootContext context, CallbackInfoReturnable<ItemStack> cir) {
-		QuiltEnchantmentHelper.clearContext();
-	}
-}
