--- library/content_other/enchantment/src/main/java/org/quiltmc/qsl/enchantment/mixin/SellEnchantedToolFactoryMixin.java
@@ -1,42 +1,0 @@
-/*
- * Copyright 2023 QuiltMC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.quiltmc.qsl.enchantment.mixin;
-
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
-
-import net.minecraft.entity.Entity;
-import net.minecraft.util.random.RandomGenerator;
-import net.minecraft.village.TradeOffer;
-
-import org.quiltmc.qsl.enchantment.api.EntityEnchantingContext;
-import org.quiltmc.qsl.enchantment.api.QuiltEnchantmentHelper;
-
-@Mixin(targets = {"net.minecraft.village.TradeOffers$SellEnchantedToolFactory"})
-public class SellEnchantedToolFactoryMixin {
-	@Inject(method = "create", at = @At(value = "INVOKE", target = "Lnet/minecraft/enchantment/EnchantmentHelper;enchant(Lnet/minecraft/util/random/RandomGenerator;Lnet/minecraft/item/ItemStack;IZ)Lnet/minecraft/item/ItemStack;"))
-	private void setEnchantingContext(Entity entity, RandomGenerator random, CallbackInfoReturnable<TradeOffer> cir) {
-		QuiltEnchantmentHelper.setContext(new EntityEnchantingContext<>(0, 0, null, entity.world, random, false, entity));
-	}
-
-	@Inject(method = "create", at = @At(value = "RETURN"))
-	private void clearEnchantingContext(Entity entity, RandomGenerator random, CallbackInfoReturnable<TradeOffer> cir) {
-		QuiltEnchantmentHelper.clearContext();
-	}
-}
