--- library/content_other/enchantment/src/testmod/resources/assets/quilt_enchantment_testmod/lang/en_us.json
@@ -1,5 +1,0 @@
-{
-  "enchantment.quilt_enchantment_testmod.reaping": "Reaping",
-  "enchantment.quilt_enchantment_testmod.pervasive": "Pervasive",
-  "enchantment.quilt_enchantment_testmod.merchant_greed": "Merchant's Greed"
-}
