+++ library/item/enchantment/src/testmod/resources/assets/quilt_enchantment_testmod/lang/en_us.json
@@ -1,0 +1,6 @@
+{
+  "item.quilt_enchantment_testmod.super_enchantable": "Super Enchantable",
+  "enchantment.quilt_enchantment_testmod.reaping": "Reaping",
+  "enchantment.quilt_enchantment_testmod.pervasive": "Pervasive",
+  "enchantment.quilt_enchantment_testmod.merchant_greed": "Merchant's Greed"
+}
